Refactor:remove user controllers to pages/api
